Revert "SWDEV-505971 - change setArgument arg from uint32_t to uint64_t"
This reverts commit 6b4e6588f79afba2a25fb43ab2773e328c91f16e.

Reason for revert: There needs to be memcpy size change



[ROCm/clr commit: 446ed661a0c60197810d630fe4b9f13192e752d4]

+++ rocclr/device/blit.hpp
@@ -236,7 +236,7 @@ class BlitManager : public amd::HeapObject {
   //! Stream batch memory operation
   virtual bool batchMemOps(const void* paramArray,
                            size_t paramSize,
-                           uint64_t count
+                           uint32_t count
   ) const = 0;
 
   //! Enables synchronization on blit operations
